requirements.txt
astropy==6.1.7
astropy-iers-data==0.2025.1.27.0.32.44
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.55.6
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyerfa==2.0.1.5
pyexiv2==2.15.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
threadpoolctl==3.5.0
tzdata==2025.1
